+++ CHANGELOG.md
@@ -4,7 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.2.0 - 2017-01-01
+## 2.0.0-0.1
 ### Changed
 
-* Release as a fork of `slog-stdlog`
+* Port to slog v2
+* Base on `slog-scope`
+
+## 1.1.0
+### Changed
+
+* BREAKING: Rewrite handling of owned values.
+
+## 1.0.1 - 2016-10-02
+### Changed
+
+* Fixed `StdLog` not serializing the key-value pairs.
+* `StdLog` message to `log` crate is lazily-evaluated.
+
+
+## 1.0.0 - 2016-09-21
+
+First stable release.